Refactor Assert.That.ExpressionEvaluation.cs into focused partial files
Splits the 581-line Assert.That.ExpressionEvaluation.cs into four focused
partial-class files, each under 300 lines, with no behavioral or API changes:

- Assert.That.ExpressionEvaluation.cs: entry points and predicate utilities
  (EvaluateExpression, RequiresSinglePassEvaluation, EvaluateAssignmentTargetSubChildren,
  IsAssignmentBinaryNodeType, IsAssignmentUnaryNodeType)
- Assert.That.ExpressionEvaluationVisitor.cs: the recursive tree visitor
  (EvaluateAllSubExpressions)
- Assert.That.ExpressionSpecialCaseEvaluation.cs: short-circuit/coalesce/ternary
  evaluators (TryEvaluateShortCircuitBinary, TryEvaluateCoalesce, TryEvaluateConditional)
- Assert.That.ExpressionRewriting.cs: subtree rebuilding
  (ReplaceSubExpressionsWithConstants, ReplaceChildWithConstant)

Fixes #9750
